Revert "fix(#449): move ingestion/extraction settings into queue window sheet…"
This reverts commit 76651c8f59c3aea8f43a09eb353ca66c98c9cde3.
